pyproject.toml
[project]
name = "lyricsmeaning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "lyricsgenius>=3.7.2",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]
